refactor(modules): correct modules structure
