+++ src/modules/Login/index.tsx
@@ -1,0 +1,1 @@
+export { LoginForm } from './components/LoginForm';
