+++ src/modules/Register/index.tsx
@@ -1,0 +1,1 @@
+export { RegisterForm } from './components/RegisterForm';
