--- src/modules/index.tsx
@@ -1,3 +1,0 @@
-export { Header } from './Header/Header';
-export { Login } from './Login/Login';
-export { Register } from './Register/Register';
